[HUDI-7970] Move methods from HoodieTableConfigUtils to HoodieTableConfig
