+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/hudi/functional/TestSevenToEightUpgrade.scala
@@ -20,8 +20,7 @@ package org.apache.hudi.functional
 
 import org.apache.hudi.DataSourceWriteOptions
 import org.apache.hudi.common.model.HoodieTableType
-import org.apache.hudi.common.table.{HoodieTableMetaClient, HoodieTableVersion}
-import org.apache.hudi.common.util.HoodieTableConfigUtils
+import org.apache.hudi.common.table.{HoodieTableConfig, HoodieTableMetaClient, HoodieTableVersion}
 import org.apache.hudi.keygen.constant.KeyGeneratorType
 import org.apache.hudi.table.upgrade.{SparkUpgradeDowngradeHelper, UpgradeDowngrade}
 import org.apache.spark.sql.SaveMode
@@ -48,7 +47,7 @@ class TestSevenToEightUpgrade extends RecordLevelIndexTestBase {
 
     // assert table version is eight and the partition fields in table config has partition type
     assertEquals(HoodieTableVersion.EIGHT, metaClient.getTableConfig.getTableVersion)
-    assertEquals(partitionFields, HoodieTableConfigUtils.getPartitionFieldPropForKeyGenerator(metaClient.getTableConfig).get())
+    assertEquals(partitionFields, HoodieTableConfig.getPartitionFieldPropForKeyGenerator(metaClient.getTableConfig).get())
 
     // downgrade table props to version seven
     // assert table version is seven and the partition fields in table config does not have partition type
@@ -56,7 +55,7 @@ class TestSevenToEightUpgrade extends RecordLevelIndexTestBase {
       .run(HoodieTableVersion.SEVEN, null)
     metaClient = HoodieTableMetaClient.reload(metaClient)
     assertEquals(HoodieTableVersion.SEVEN, metaClient.getTableConfig.getTableVersion)
-    assertEquals("partition", HoodieTableConfigUtils.getPartitionFieldPropForKeyGenerator(metaClient.getTableConfig).get())
+    assertEquals("partition", HoodieTableConfig.getPartitionFieldPropForKeyGenerator(metaClient.getTableConfig).get())
 
     // auto upgrade the table
     // assert table version is eight and the partition fields in table config has partition type
@@ -66,6 +65,6 @@ class TestSevenToEightUpgrade extends RecordLevelIndexTestBase {
       validate = false)
     metaClient = HoodieTableMetaClient.reload(metaClient)
     assertEquals(HoodieTableVersion.EIGHT, metaClient.getTableConfig.getTableVersion)
-    assertEquals(partitionFields, HoodieTableConfigUtils.getPartitionFieldPropForKeyGenerator(metaClient.getTableConfig).get())
+    assertEquals(partitionFields, HoodieTableConfig.getPartitionFieldPropForKeyGenerator(metaClient.getTableConfig).get())
   }
 }
